la folie

+++ source/states/PlayState.hx
@@ -371,14 +371,13 @@ class PlayState extends FlxState
 		//debugCanvas.drawCircle(FlxG.mouse.x, FlxG.mouse.y, 6, FlxColor.TRANSPARENT, { color: FlxColor.RED, thickness: 2 });
 		#end
 		
-		//for (target in targets) {
-			//target.body.angularVel = FlxAngle.asRadians(instantRotation * 60);
-		//}
+		for (target in targets) {
+			target.body.angularVel = FlxAngle.asRadians(instantRotation * 60);
+		}
 		
-		//for (targeta in targetsHitarea) {
-			////targeta.angularVelocity = FlxAngle.asRadians(instantRotation * 60);
-			////targeta.angle= FlxAngle.asRadians(instantRotation * 60);
-		//}
+		for (targetHitArea in targetsHitarea) {
+			targetHitArea.angularVelocity = instantRotation * 60;
+		}
 		
 		if (targets.length == 0 && !gameEnd) {
 			trace("You win");
@@ -417,46 +416,29 @@ class PlayState extends FlxState
 	public function onBulletCollides(callback:InteractionCallback) {
 		trace(callback.int2.userData.id);
 		
-		if(callback.int2.userData.type !=null)
-		{
-			var body = callback.int1.castBody;
-			var body2 = callback.int2.castBody;
+		if(callback.int2.userData.type != null) {
+			var projectileBody = callback.int1.castBody;
+			var targetBody = callback.int2.castBody;
 			
-			//var initialVector : FlxVector = FlxVector.get(body.position.x, body.position.y);
+			var centerProjectileP : FlxPoint = new FlxPoint(projectileBody.position.x, projectileBody.position.y);
+			var centerTargetP : FlxPoint = new FlxPoint(targetBody.position.x, targetBody.position.y);
 			
-			var initialRot = callback.int2.userData.parent.initialRotation;
-			trace("INIT : " + initialRot);
+			var targetRotation = FlxAngle.asDegrees(targetBody.rotation);
+			if (targetRotation < 0) {
+				targetRotation += 360;
+			}
+			var vectorTargetToProjectile = FlxVector.get(centerProjectileP.x - centerTargetP.x, centerProjectileP.y - centerTargetP.y);
+			var vectorTargetLookingAt = FlxVector.get(0, 1).rotateByDegrees(targetRotation);
 			
-			var x2 = FlxMath.fastCos(0) - Math.sin(Std.parseInt(initialRot) * 1);
-			var y2 = Math.sin( 0) + FlxMath.fastCos(Std.parseInt(initialRot) * 1);
+			var collisionAngle = vectorTargetLookingAt.degreesBetween(vectorTargetToProjectile);
 			
-
-			
-			var centerProjectileP : FlxPoint = new FlxPoint(body.position.x, body.position.y);
-			var centerTargetP : FlxPoint = new FlxPoint(body2.position.x, body2.position.y);
-			
-			var vectorTarget =  new FlxVector(centerTargetP.x - centerProjectileP.x, centerTargetP.y - centerProjectileP.y);
-			var vectorInitial =  new FlxVector(x2,y2);
-			
-			var angle = FlxMath.dotProduct(vectorTarget.x, vectorTarget.y, vectorInitial.x, vectorInitial.y);
-			trace("ANGLE : " + angle);
-			
-			
-			
-			if (callback.int2.userData.type == TargetType.FIXED)
-			{
-				//if (Math.abs(angle) > 10)
-				//{
+			if (callback.int2.userData.type == TargetType.FIXED) {
+				if (collisionAngle > 150 && collisionAngle < 210) {
 					callback.int2.userData.parent.kill();
 					targets.remove(callback.int2.userData.parent, true);
-				//}
-				
-				
+				}
 			}
-
-			//trace(callback.int2.userData.parent);
-			//trace($type(callback.int2.cbTypes));
-			//trace($type(callback.int2.cbTypes));
+			
 			projectile.body.velocity.setxy(0, 0);
 			projectile.state = ON_TARGET;
 		}
